refactor: save army copy hook
