+++ modules/armies/hooks/index.ts
@@ -1,1 +1,2 @@
+export { default as useSaveArmyCopy } from './useSaveArmyCopy'
 export { default as useUnitTypes } from './useUnitTypes'
